Remove room while getting

+++ api/get.js
@@ -3,6 +3,6 @@ import connectDB, {client as mongo} from "../db";
 export default async (req, res) => {
     await connectDB()
     const room = await mongo.db("chat").collection("queue").findOne()
-    await mongo.db("chat").collection("queue").deleteOne(room)
+    if (room) await mongo.db("chat").collection("queue").deleteOne(room)
     return res.json({room})
 }
